requirements.txt
pandas>=1.5.0
scikit-learn>=1.0.0
pyyaml>=6.0
joblib>=1.2.0
dvc[ssh]>=3.0.0